feat: if config key darknet.shutup is false, no stdio redirection is done
This can be necessary if other library is storing stdout/stderr file
descriptors (eg for logging), and our little trick makes them try to
write to a closed stream. To suppress darknet verbosity in this case,
there is no option but to define printf and fprintf as empty in darknet
and recompile it.